fix(ci): unignore apps/web/lib + satisfy ruff on Phase 1.1 files
The root .gitignore's unanchored 'lib/' rule (from the Python distutils
template) was matching apps/web/lib/, silently dropping the three TS
modules that ChatPanel / ChatTurn / CoverageRefusal / SchemaTour all
import from @/lib/{types,api}. CI on commit 113a397 reported 8x
'Cannot find module' + 2x implicit-any (downstream of the missing
StreamHandler type) plus 8 ruff violations in the new Phase 1.1 code.

Changes:
  - /lib/, /lib64/ are now anchored to repo root so the frontend's
    apps/web/lib/ stays tracked.
  - Add apps/web/lib/{types,api,api-types}.ts (previously ignored).
  - explore.py: drop unused AIMessage import.
  - prompts.py: collapse the stray double blank line after
    __future__ import (I001).
  - profiler.py: drop redundant int() around round() — round(x)
    already returns int in Python 3 (RUF046 x2).
  - tests/test_{coverage,explore,profiler}.py: ruff --fix import
    blocks; drop unused pytest import in test_profiler.

Verified locally on a clean HEAD worktree:
  backend  : ruff ok, mypy ok (50 files), pytest -m 'not integration' ok (365)
  frontend : typecheck ok, vitest ok (20/20), next build ok

+++ apps/api/copilot/profiler.py
@@ -164,7 +164,7 @@ def _column_stats(eng: Engine, table: str) -> dict[str, dict[str, Any]]:
             distinct_count = int(n_distinct)
         else:
             # Negative => fraction of the table. Multiply by row count.
-            distinct_count = int(round(abs(float(n_distinct)) * row_count)) if row_count else None
+            distinct_count = round(abs(float(n_distinct)) * row_count) if row_count else None
 
         sample_values = _parse_pg_array(mcv_text)[:_SAMPLE_VALUES_LIMIT] if mcv_text else None
         hist = _parse_pg_array(hist_text) if hist_text else None
@@ -518,7 +518,7 @@ def _format_column_line(col: dict[str, Any]) -> str:
     if ndv is not None:
         tags.append(f"{ndv} distinct")
     if null_ratio is not None:
-        pct = int(round(null_ratio * 100))
+        pct = round(null_ratio * 100)
         if pct > 0:
             tags.append(f"{pct}% null")
     if samples:
